make: Support building without Cask
When WITHOUT_CASK is non-empty, then don't use Cask.

The main advantage of not using Cask is that this approach compiles
each library in a separate Emacs instance, which allows making sure
that all dependencies are properly defined.

This assumes that each dependency is installed at "../<PACKAGE>",
and the user is responsible for setting that up.  So we keep the
Cask variant as well.

The "test" target still always relies on Cask.