Remove "-face" suffix from names of faces
As mentioned at (info "(emacs)Defining Faces") "[the name of a face]
should not end in "-face" (that would be redundant).  For historic
reasons a few built-in faces actually do end with "-face" and it so
happens that many of our faces are closely related to just those
`font-lock' and `compilation' faces and that probably inspired our
use of the suffix.  Even so, we should stop now.