Don't needlessly depend on async package
The purpose of the `async' package is to let you easily run elisp
code in an inferior emacs instance.  This is not how we use it.

While `async' also has some support for running other processes
asynchronously, one should not depend on this package solely to
do this sort of thing.  Emacs provides support for doing that
itself.

Let's use that.  The new `rustic-doc--start-process' fatefully
reproduces the implementation in `async' to the extend necessary
to not change the behavior here.