Improve how essential libraries are required
A few essential libraries, which provide language features that we
(and many others) heavily rely on are not part of Emacs' official
"prelude" (in the Rust sense).

Instead of trying to explicitly requiring these libraries in each and
every of our libraries that used them (and failing to do so), we load
them explicitly just once, in "rustic-common", which directly or
indirectly is required by all of our other libraries.

We extend the prelude with the built-in `cl-lib', `pcase', `seq' and
`subr-x', and the third-party `dash'.

We are not doing the same for the third-party `s' and `f', which IMO
it would be better to not use those at all, so I would rather contain
their use to the libraries that unfortunately already use them.