Avoid loading rustic-common directly
"rustic-common.el" defines an assorted set of things that have to be
defined early on.  We could merge this library back into "rustic.el"
but do not do so to make it easier to compare that with "rust-mode.el".

`rustic' requires `rustic-common' and all other libraries depend on
the former, whether explicitly or indirectly.  Since the existence
of `rustic-common' is a bit of an implementation detail, we stop
depending on that in all libraries, except in `rustic' itself.