Require version 0.3.0 of "project" package
This package is distributed with Emacs but it is also available on GNU
Elpa to make it possible to install the latest version on older Emacs
releases.

Since version 0.3.0 `project-roots' has been deprecated in favor of
`project-root':

> rustic-rustfmt.el:198:25: Warning: ‘project-roots’ is an obsolete
>     function (as of 0.3.0); use ‘project-root’ instead.

The best way to deal with that is to use `project-root' and make
users update to a "project" version that features that function.