test: Support compiling test files
It appears that ert-runner does not support compiling test files, so
we only do it when not using Cask and ert-runner.  Unfortunately that
means that most contributors won't benefit for the time being.

Ad: When using the `auto-compile' package, then saving an elisp file
causes that to be automatically recompiled.  This means that one can
note bugs as one introduces them, which is something Rust developer
can surely appreciate.

This does not work so well for the test files because they cannot
require `test-helper', because "test-helper.el" does not, and should
not, provide a feature.  Non-libraries, such as test files should
never provide features.  It is acceptable for a test files to provide
a feature, iff used the package prefix, but it is still better to
avoid doing so.  If "test-helper.el" specifically provided a feature,
then that would be a big mistake as that would conflict with the files
by the same name that come with the many other packages that use
ert-runner.  (In fact back in the days hundreds of them did provide
the feature and it took me a long time to fix that.)

We bypass all these complications by dropping a hint in the test
files, so that one can manually, yet quickly, load "test-helper.el"
when needed.  That isn't perfectly nice, but its a lot better than
the alternatives.