Made cargo-run-current-test choose closest of mod or fn to run.
Previously, this function would always pick the previous fn to run,
even when the current point was a module or inside a module. The
function could be from a previous module, which didn't make sense.
Running all tests in a module is a common thing to do.
This change allows users to do that.
If they are within a fn, the fn only is run.
If they are not within a fn but there's a surrounding mod, then the
whole mod is run.