use polymode for interaction in rustic-cargo-run
rustic-cargo-run-mode is renamed to rustic-cargo-plain-run-mode.
rustic-cargo-comint-run-mode is added anew.  rustic-cargo-run-mode now
refers to a wrapper mode that switches between the two depending on
the setting.

Known issue: once the user interacts with the compiled program, this
creates an indirect buffer named *rustic-compilation*[comint], which
subsequent calls to rustic-cargo-run doesn't clean up.  So buffers
named like this will accumulate.