Refactor and improve the ability to do cargo run commands
Summary of the changes:

- We have two modes now: rustic-cargo-run-mode and
rustic-cargo-run-comint-mode. The first one is primarily meant for
interactive usage and the comint mode is used for passing data to it.
- Remove the dependency on polymode. The whole thing felt hacky
IMO. If you want to have highlighting on comile errors, the
recommendation is to just use `rustic-compile` before going to the run
phase.
- We just pass the data directly in the comint buffer now instead of
using `rustic-compile-send-input`. We have removed that function now
as it's not required anymore.