Fixes a bug when dependencies are specified as git repositories
The `cargo outdated --depth 1` doesn't always gives the package from
line number 2. For a sample project I'm having this is the output it
gives:

``` shellsession
❯ cargo outdated --depth 1
    Updating git repository `https://github.com/snoyberg/async-refresh-rs`
Name           Project  Compat   Latest   Kind    Platform
----           -------  ------   ------   ----    --------
anyhow         1.0.40   1.0.53   1.0.53   Normal  ---
askama         0.10.5   ---      0.11.0   Normal  ---
```

As you can see in the above case, the hardcoding of the value 2
doesn't work. The above scenario, makes cargo outdated stuff to break
entirely. Infact, for me the buffer was empty and the sentinel
function broke. The fix instead tries to check for `--` character to
delimit where the packages are starting from.