Fix cargo upgrade when you update version to latest version
Right now, rustic-cargo-upgrade-execute only works when we try to
upgrade it to a compabible version ie

``` shellsession
Name           Project  Compat   Latest   Kind    Platform
----           -------  ------   ------   ----    --------
anyhow         1.0.40   1.0.53   1.0.53   Normal  ---
```

Trying to upgrade above will work. But this will fail:

``` shellsession
Name           Project  Compat   Latest   Kind    Platform
----           -------  ------   ------   ----    --------
askama         0.10.5   ---      0.11.0   Normal  ---
```

And the reason is that it tries to use this code `(elt crate 2)` for
taking the version from. But this makes it to work only when the
`compat` column is already filled. Also because of the way
`tabulated-list` works, we just change the buffer when we do
`rustic-cargo-mark-upgrade`. This commit fixes it in such a way that
it works for all the cases.