Remove testing actual buffer contents
This is mostly because we don't have cargo expand in the CI system and
no binary seems to be available. Increasing the CI time by ~4 minutes,
just for one test doesn't seem good to me. So the test just checks if
the proper command is executed and that is good enough IMO.