Use `default-directory` in popup buffer instead of `buffer-file-name`
`buffer-file-name` is not defined in the rustic popup buffer (which has focus after it is popped up)
since it isn't associated with a buffer, but `default-directory` does.

Since `rustic-cargo-run-get-relative-example-name` is only looking at the first component of the
relative path and checking to see if it is in the examples directory, the full path to the file
buffer does not matter.

fixes #392