Use inheritenv wherever an auxiliary buffer is created, if available
In order for process-local environments to be propagated to auxiliary buffers, it is
necessary to wrap buffer-creating code in inheritenv, as per
https://github.com/purcell/inheritenv

An example requirement for this is with Nix development shells via
direnv, as per
https://nixos.wiki/wiki/Development_environment_with_nix-shell

This change adds inheritenv wherever one of these is called:

- get-buffer-create
- with-temp-buffer

except for places where the latter has only trivial scope and does
nothing interesting.

Subsequent additions of code which creates such auxiliary buffers
should do likewise.

Add rustic--with-temp-process-buffer for environment propagation

Workaround for with-temp-buffer not propagating the environment, as per
https://github.com/magit/magit/pull/4169

Also propagate exec-path in with-temp-buffer workaround.

Load lsp with lsp-deferred so happens after envrc

Remove some unnecessary inheritenv

Not needed around the new with-temp-buffer wrapper

Also inheritenv for generate-new-buffer in rust-doc

Fix indent

Only use inheritenv if available.