Fix org-babel workflow when auto wrapping main function
This bug is easily reproducible when you use try something like this:

```
let x = [1,2,3,4,5,6];
println!("{x:?}");
```

And then try executing it. While you will get the result properly, you
could see in the Messages buffer that that would be some errors
related to compilation buffer.

And the reason is this: It tries to do rustfmt on the babel block and
fails. In this fix, I just disable doing rustfmt when we have enabled
automatic wrapping or when we instruct it to wrap the main function.

Another alternative fix could be do rustfmt on the wrapped main
function and the remove the wrapped main function and render it again
on the babel blocks. But IMO, that's going to be very fragile.