Fix bug when only one crate needs updates
I found a bug when using cargo update that I introduced with my last
refactor. The issue is that I was trying to check whether we had a
cargo project without a workspace. In this case the goal was to append
a dummy name ("---"). Unfortunately, I only checked how many crates in
the workspace there where, not whether the first crate doesn't have a
name (this only happens in projects without a workspace)