Make `rustic-format-start-process' work over tramp.
`rustic-rustfmt-bin' selects the rustfmt binary to run depending on
whether the current buffer is local (in which case
`rustic-rustfmt-bin' will be used) or remote
(`rustic-rustfmt-bin-remote').

`rustic-format-start-process' invokes `rustic-rustfmt-bin' inside a
`with-current-buffer' form that makes the error buffer current, which
will always seem local, regardless of whether the buffer being
formatted is local or not

This patch invokes `rustic-rustfmt-bin' as one of the let bindings
so that the correct binary will be selected.