Correctly handle directory on rerun
Currently rustic will override the default-directory to the workspace root after
a compilation is finished. This enables it to correctly find the paths in error
messages. But it also means that if you rerun a command it will rerun in the
root crate instead of the original one.

This change adds a new variable to track the original directory and resets it
when we rerun.